requirements.txt
anchorpy==0.18.0
anchorpy-core==0.1.3
anyio==3.6.2
appnope==0.1.3
asttokens==2.1.0
attrs==22.1.0
autoflake==1.7.7
backcall==0.2.0
based58==0.1.1
black==22.10.0
borsh-construct==0.1.0
cachetools==4.2.4
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0ek9yztvu3/croot/certifi_1665076692562/work/certifi
charset-normalizer==2.1.1
click==8.1.3
construct==2.10.68
construct-typing==0.5.3
decorator==5.1.1
exceptiongroup==1.2.0
executing==1.2.0
genpy==2021.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
iniconfig==1.1.1
ipython==8.6.0
jedi==0.18.1
jsonalias==0.1.1
jsonrpcclient==4.0.2
matplotlib-inline==0.1.6
more-itertools==8.14.0
mypy-extensions==0.4.3
numpy==1.23.4
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.3
pluggy==1.0.0
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyflakes==2.5.0
Pygments==2.13.0
pyheck==0.1.5
pyparsing==3.0.9
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-xprocess==0.18.1
python-dotenv==0.21.0
pytools==2022.1.12
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
solana==0.30.2
solders==0.18.1
stack-data==0.6.0
sumtypes==0.1a6
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
traitlets==5.5.0
typer==0.4.1
types-cachetools==4.2.10
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
websockets==10.4
zstandard==0.18.0
